+++ src/main/java/com/project/team4backend/domain/auth/contoller/AuthController.java
@@ -1,7 +1,7 @@
 package com.project.team4backend.domain.auth.contoller;
 
 import com.project.team4backend.domain.auth.dto.response.AuthResDTO;
-import com.project.team4backend.domain.auth.service.command.AuthCommandService;
+import com.project.team4backend.domain.auth.service.command.auth.AuthCommandService;
 import com.project.team4backend.global.apiPayload.CustomResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
