+++ src/main/java/com/project/team4backend/global/security/filter/CustomLoginFilter.java
@@ -7,8 +7,8 @@ import com.project.team4backend.global.apiPayload.CustomResponse;
 import com.project.team4backend.global.security.CustomUserDetails;
 import com.project.team4backend.global.security.JwtUtil;
 import com.project.team4backend.domain.auth.dto.response.AuthResDTO.JwtResDTO;
-import com.project.team4backend.domain.auth.exception.AuthErrorCode;
-import com.project.team4backend.domain.auth.exception.AuthException;
+import com.project.team4backend.domain.auth.exception.auth.AuthErrorCode;
+import com.project.team4backend.domain.auth.exception.auth.AuthException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
